+++ tiledb/sm/misc/constants.h
@@ -503,6 +503,9 @@ extern const std::string redirection_header_key;
 /** List of possible certificates files for libcurl */
 extern const std::array<std::string, 6> cert_files_linux;
 #endif
+
+/** Delimiter for lists passed as config parameter */
+extern const std::string config_delimiter;
 }  // namespace constants
 
 }  // namespace sm